Fixed IE8 bug where slide area in banner appeared as a white box.
I'm not entirely sure what cleartype fixing for IE is, but
I noticed that when I removed the call to $.cycle, the problem went away,
and then I noticed there was an option in the plugin described as follows:

"""
set to true to disable extra cleartype fixing (leave false to force
background color setting on slides)
"""

That seemed like it could be promising, so I changed the value and
things worked. I'm still not entirely sure *why* they work because I
don't know what cleartype fixing is, but it appears to only apply to IE,
and other browsers don't break when this setting is changed, so it looks
like a decent fix.